[JXD-231] Более гибкий подход, когда каждый рендер в своей папке
