+++ esphome/components/display_menu_render_dallas/display_menu_render_dallas.cpp
@@ -1,36 +1,10 @@
-#include "display_menu_renderers.h"
+#include "display_menu_render_dallas.h"
+
+#include "esphome/components/display_menu_render_sensor/display_menu_render_sensor.h"
 
 namespace esphome {
 namespace display_menu_renderers {
 
-#ifdef USE_SWITCH
-size_t SwitchMenuRender::render_entity(MenuItemMenu *menu, const groups::EntityInfo &info) {
-  switch_::Switch *switch_obj = static_cast<switch_::Switch *>(info.entity);
-  MenuItemSwitch *n_switch = new MenuItemSwitch();
-  n_switch->set_text(switch_obj->get_name());
-  n_switch->set_immediate_edit(true);
-  n_switch->set_switch_variable(switch_obj);
-  n_switch->set_on_text("On");
-  n_switch->set_off_text("Off");
-
-  menu->add_generated_items(n_switch);
-  return 1;
-}
-#endif
-
-#ifdef USE_SENSOR
-size_t SensorMenuRender::render_entity(MenuItemMenu *menu, const groups::EntityInfo &info) {
-  sensor::Sensor *sensor_obj = static_cast<sensor::Sensor *>(info.entity);
-  MenuItem *item = new MenuItem(MENU_ITEM_LABEL);
-  auto lambda = get_render_lambda(sensor_obj, true);
-  item->set_text(lambda);
-
-  menu->add_generated_items(item);
-  return 1;
-}
-#endif
-
-#ifdef USE_DALLAS
 size_t DallasTempMenuRender::render_entity(MenuItemMenu *menu, const groups::EntityInfo &info) {
   dallas_temp::DallasTemperatureSensor *sensor_obj = static_cast<dallas_temp::DallasTemperatureSensor *>(info.entity);
 
@@ -65,7 +39,6 @@ void DallasTempMenuRender::proccess_submenu(MenuItemMenu *menu, dallas_temp::Dal
   menu->add_item(temp_info);
   menu->add_item(address_info);
 }
-#endif
 
 }  // namespace display_menu_renderers
 }  // namespace esphome
